Update various docstrings, add additional error handling, made typehinting consistent, add TO-DO's
Various docstrings have been updated to give additional or updated information.

Some functions have received overall better error handling. There is now a helper function that takes the function's name and appends a message to it. This may be largely unnecessary, but I like it.

Typehinting is now overall more consistent in spacing and applications.

A couple of TO-DO's have been added (and moved around).