Implement new Comparer class
Can be extended a bit more for some more exotic comparisons but the ones
there currently should cover most uses (especially since the most useful
comparisons can now be done with vspreview directly).

Simplifies the actual compare logic of a few existing functions and adds
a few new features including being able to specify names for the clips
and a direction for stacks.