scale: descale: output float clip
This is done to make sure the mask functions properly. At some point I plan on fixing it so it accurately scales the mask's threshold if a) default mask is passed, and b) the input clip isn't float. But for now, this is a stop-gap solution so people (*cough cough* eX) won't yell at me when they're being stupid