deps: fix stgfunc branch reference
stgfunc's default branch is `master`, not `main`.

This should have been caught by CI, and sure enough there is an error there:

    Collecting stgfunc@ git+https://github.com/Jaded-Encoding-Thaumaturgy/stgfunc.git@main (from -r requirements.txt (line 10))
      Cloning https://github.com/Jaded-Encoding-Thaumaturgy/stgfunc.git (to revision main) to c:\users\runneradmin\appdata\local\temp\pip-install-gfqtr21d\stgfunc_40b67a68613b40cf9ab977f3f2ccaada
      Running command git clone --filter=blob:none --quiet https://github.com/Jaded-Encoding-Thaumaturgy/stgfunc.git 'C:\Users\runneradmin\AppData\Local\Temp\pip-install-gfqtr21d\stgfunc_40b67a68613b40cf9ab977f3f2ccaada'
      WARNING: Did not find branch or tag 'main', assuming revision or ref.
      Running command git checkout -q main
      error: pathspec 'main' did not match any file(s) known to git
      error: subprocess-exited-with-error

But looks like this wasn't causing the step to fail. From a Stack Overflow post
(<https://stackoverflow.com/a/75835600>) it looks like on Windows agents we
gotta set the default shell to bash in order to fail on errors in multi-line run
steps like the dependencies one. So I included that change here to hopefully
catch stuff like this in the future.

Also updated third-party Actions while I was at it and removed the conditional
on the mypy step so that it actually runs (again the conditional was referencing
a step id that didn't exist plus it looked superfluous).