get_match_centers_scaling: Fix incorrect docstring example
Bit uglier to look at, but it now shows an accurate depiction of the match edges vs. centers models.

The previous example was wrong because it did not treat the pixels as "samples" correctly (the center point of the pixels on the far left/right didn't match correctly in the match centers model).